pyproject.toml
[project]
name = "ledger_ai_assistant"
version = "0.1.0"
description = "..."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "jq>=1.10.0",
    "langchain-anthropic>=0.3.19",
    "langchain-aws>=0.2.31",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "langchain-groq>=0.3.7",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "langmem>=0.0.29",
    "matplotlib>=3.10.6",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.2",
    "sentence-transformers>=5.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
